Add Date command
This command behaves in two ways.
When only number is given, it shows the two days that's behind and
ahead for the given number.
When number and date (such as 2023-01-01) are given, the base date
for the calculation is the given date and it also shows whether
today is in the date range.